sync

+++ rust/gas_with_wasm/src/lib.rs
@@ -1,2 +1,2 @@
 #[no_mangle]
-pub fn add(left: u64, right: u64,) -> u64 { left + right }
+pub fn add(left: i32, right: i32,) -> i32 { left + right }
